Modify the gtk frontend to read X input from stdin.
The code in xdo-stdin.c creates artificial X mouse and key events
according to stdin. Each command is a single character with parameters
as follows:

  c%5  Parameters           Description
  -----------------------------------------------------
  0    x:uint16 y:uint16    Move mouse
  1    letter:uint8         Press and release a key
  2    button:uint8         Mouse click
  3    button:uint8         Mouse button down
  4    button:uint8         Mouse button up

By default, the events are created on hardcoded display :1, because it
was too easy to accidentally test on a production display. Change it
in xdo-stdin.c.

Up to 200 events are read and generated this way, then the program
exits (tunable in xdo-stdin.c).

The puzzle code has also been modified to ignore load/save operations.